requirements.txt
fastapi
uvicorn
python-dotenv
requests
pypdf
python-multipart

